Add data/db/.gitkeep as workaround for user-namespaces
When you use docker with user-namespaces and mount volumes which don't
exist, docker will create these volumes as root.  The problem with this
approach is that the container will not have the right access rights to
write to the root-owned folder.  This causes problems when launching the
docker-compose in the previous setting.  If we clone the repository and
make sure the folder exists, virtuoso will happily start up in its
container.