Ignore Virtuoso data files
You shouldn't store your Virtuoso data in the commits, you should backup
the data in triple-form and store them separately in most cases.  We are
ignoring the Virtuoso content to make this clear.  We need to explicitly
ignore the current files as there may be content in toLoad for
automatic loading on setup.